IS-1663: Validere tilstandsendringer
