+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/domain/Aktivitetskrav.kt
@@ -13,19 +13,6 @@ import java.util.*
 
 const val AKTIVITETSKRAV_STOPPUNKT_WEEKS = 8L
 
-enum class AktivitetskravStatus {
-    NY,
-    AVVENT,
-    UNNTAK,
-    OPPFYLT,
-    AUTOMATISK_OPPFYLT,
-    FORHANDSVARSEL,
-    STANS,
-    IKKE_OPPFYLT,
-    IKKE_AKTUELL,
-    LUKKET,
-}
-
 data class Aktivitetskrav(
     val uuid: UUID,
     val personIdent: PersonIdent,
